+++ NEWS.md
@@ -10,6 +10,7 @@ New features(CLI,Configs)
 New features(Analysis)
 + Add function signatures for functions added/modified in PHP 7.3. (#1537)
 + Improve the line number for warnings about unextractable `@property*` annotations.
++ Make Phan aware that `$x` is not false inside of loops such as `while ($x = dynamic_value())` (#1646)
 
 Plugins:
 + Add `UnknownElementTypePlugin` (currently a work in progress).